+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [patch]
 
-- Added `RecoverableAccount` contract, an all-in-one EIP-7702 account with ERC-7821 batching, ERC-4337 gas sponsorship, and ERC-7947 recovery.
+- Added `RecoverableAccount` contract, an all-in-one EIP-7702/ERC-4337 account with ERC-7821 batching, ERC-4337 gas sponsorship, and ERC-7947 recovery.
 - Added eslint.
 
 ## [3.2.3]